solve some comments changing the sliding window radius to double

+++ src/planning/include/config/path_calculation_config.hpp
@@ -21,7 +21,7 @@ struct PathCalculationConfig {
   int max_points_ = 50;
   int reset_global_path_ = 10;
   int skidpad_minimum_cones_ = 10.0;
-  int sliding_window_radius_ = 20;
+  double sliding_window_radius_ = 20.0;
   bool use_sliding_window_ = true;
 
   PathCalculationConfig() = default;
